+++ CHANGELOG.md
@@ -8,10 +8,16 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
-*   add SquareException.toObject() method ([d6e194d](https://github.com/goparrot/square-connect-plus/commit/d6e194dc664b6c94b67f0e3c0d40645b71cf1bd4))
+-   add SquareException.toObject() method ([d6e194d](https://github.com/goparrot/square-connect-plus/commit/d6e194dc664b6c94b67f0e3c0d40645b71cf1bd4))
 
 ### 0.0.1 (2020-01-31)
 
 ### Features
 
-*   **client:** add client with retry mechanism ([ca771bd](https://github.com/goparrot/square-connect-plus/commit/ca771bd74bf0d39f012566d2c19d9af24d030555))
+-   **client:** add client with retry mechanism ([ca771bd](https://github.com/goparrot/square-connect-plus/commit/ca771bd74bf0d39f012566d2c19d9af24d030555))
+
+### 1.0.0 (2022-01-20)
+
+### Features
+
+-   migrate to Square Node.js SDK([c1de3e1](https://github.com/goparrot/square-connect-plus/pull/29/commits/c1de3e10c339a45e8bf4ad05412b980b09677eb0))